requirements.txt
crewai
futureagi
requests
python-dotenv
